# lib/api-zod/src/generated/types/compatibilityInviteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompatibilityBirthInput } from './compatibilityBirthInput';

export interface CompatibilityInviteInput {
  /**
     * Ignored for ownership
     * @deprecated
     */
  createdByUserId?: string;
  relationshipType?: string;
  birthProfile: CompatibilityBirthInput;
}